refactor: remove foundry config and linking crates
