+++ Cargo.toml
@@ -12,7 +12,6 @@ members = [
     "crates/foundry/cheatcodes",
     "crates/foundry/cheatcodes/spec",
     "crates/foundry/common",
-    "crates/foundry/config",
     "crates/foundry/evm/core",
     "crates/foundry/evm/coverage",
     "crates/foundry/evm/evm",
@@ -158,6 +157,7 @@ wildcard-imports = "warn"
 zero_sized_map_values = "warn"
 
 [workspace.dependencies]
+edr_defaults = { path = "crates/edr_defaults" }
 edr_test_utils = { path = "crates/edr_test_utils" }
 
 forge = { path = "crates/foundry/forge" }
@@ -166,14 +166,12 @@ foundry-cheatcodes = { path = "crates/foundry/cheatcodes" }
 foundry-cheatcodes-spec = { path = "crates/foundry/cheatcodes/spec" }
 foundry-cli = { path = "crates/foundry/cli" }
 foundry-common = { path = "crates/foundry/common" }
-foundry-config = { path = "crates/foundry/config" }
 foundry-evm = { path = "crates/foundry/evm/evm" }
 foundry-evm-core = { path = "crates/foundry/evm/core" }
 foundry-evm-coverage = { path = "crates/foundry/evm/coverage" }
 foundry-evm-fuzz = { path = "crates/foundry/evm/fuzz" }
 foundry-evm-traces = { path = "crates/foundry/evm/traces" }
 foundry-macros = { path = "crates/foundry/macros" }
-foundry-linking = { path = "crates/foundry/linking" }
 
 # solc & compilation utilities
 foundry-block-explorers = { version = "=0.2.7", default-features = false }
