+++ crates/edr_defaults/Cargo.toml
@@ -4,6 +4,7 @@ version.workspace = true
 edition.workspace = true
 
 [dependencies]
+alloy-primitives.workspace = true
 
 [lints]
 workspace = true
