--- crates/foundry/config/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "foundry-config"
-description = "Foundry configuration"
-
-version.workspace = true
-edition.workspace = true
-
-[dependencies]
-foundry-compilers = { workspace = true, features = ["svm-solc"] }
-
-alloy-chains = { workspace = true, features = ["serde"] }
-alloy-primitives = { workspace = true, features = ["serde"] }
-
-dunce.workspace = true
-eyre.workspace = true
-Inflector = "0.11"
-number_prefix = "0.4"
-regex = "1"
-semver = { version = "1", features = ["serde"] }
-tracing.workspace = true
-
-[target.'cfg(target_os = "windows")'.dependencies]
-path-slash = "0.2.1"
-
-[dev-dependencies]
-similar-asserts.workspace = true
-tempfile.workspace = true
-
-[lints]
-workspace = true
