+++ crates/foundry/evm/evm/Cargo.toml
@@ -8,7 +8,6 @@ edition.workspace = true
 [dependencies]
 foundry-cheatcodes.workspace = true
 foundry-common.workspace = true
-foundry-config.workspace = true
 foundry-evm-core.workspace = true
 foundry-evm-coverage.workspace = true
 foundry-evm-fuzz.workspace = true
