+++ crates/foundry/evm/fuzz/Cargo.toml
@@ -7,7 +7,6 @@ edition.workspace = true
 
 [dependencies]
 foundry-common.workspace = true
-foundry-config.workspace = true
 foundry-evm-core.workspace = true
 foundry-evm-coverage.workspace = true
 foundry-evm-traces.workspace = true
