+++ crates/foundry/evm/fuzz/src/fuzz_config.rs
@@ -1,5 +1,3 @@
-//! Configuration for fuzz testing.
-
 use std::path::PathBuf;
 
 use alloy_primitives::U256;
