+++ crates/foundry/forge/tests/it/core.rs
@@ -6,7 +6,7 @@ use edr_test_utils::SolidityTestFilter;
 use forge::result::SuiteResult;
 use foundry_evm::traces::TraceKind;
 
-use crate::{config::*, test_helpers::TEST_DATA_DEFAULT};
+use crate::{config::*, helpers::TEST_DATA_DEFAULT};
 
 #[tokio::test(flavor = "multi_thread")]
 async fn test_core() {
