+++ crates/foundry/forge/tests/it/spec.rs
@@ -3,7 +3,7 @@
 use edr_test_utils::SolidityTestFilter;
 use foundry_evm::revm::primitives::SpecId;
 
-use crate::{config::*, test_helpers::TEST_DATA_DEFAULT};
+use crate::{config::*, helpers::TEST_DATA_DEFAULT};
 
 #[tokio::test(flavor = "multi_thread")]
 async fn test_shanghai_compat() {
