--- crates/foundry/linking/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "foundry-linking"
-description = "Smart contract linking tools"
-
-version.workspace = true
-edition.workspace = true
-
-[dependencies]
-foundry-compilers = { workspace = true, features = ["full"] }
-semver = "1"
-alloy-primitives = { workspace = true, features = ["rlp"] }
-thiserror = "1"
-
-[lints]
-workspace = true
